+++ src/components/popup-card/CardModal.js
@@ -6,10 +6,10 @@ import CardDetails from "./CardDetails";
 import CardCategories from "./CardCategories";
 import CardSection from "./CardSection";
 import CardGameName from "./CardGameName";
-import AddButtonCollection from "../AddButtonCollection";
+import AddButtonCollection from "./AddButtonCollection";
 import Star from "../../icons/Star";
 import Dice from "../../icons/Dice";
-import AddButtonWishlist from "../AddButtonWishlist";
+import AddButtonWishlist from "./AddButtonWishlist";
 import { postGameToCollection } from "../../api/GameCollection";
 import { postGameToWishlist } from "../../api/WishlistCollection";
 import DetailButton from "../DetailButton";
